+++ packages/terra-site/src/examples/date-picker/Index.jsx
@@ -13,6 +13,7 @@ import DatePickerIncludeDates from './DatePickerIncludeDates';
 import DatePickerFilterDates from './DatePickerFilterDates';
 import DatePickerMinMax from './DatePickerMinMax';
 import DatePickerDefaultDate from './DatePickerDefaultDate';
+import DatePickerInsideModal from './DatePickerInsideModal';
 
 const DatePickerExamples = () => (
   <div>
@@ -41,6 +42,11 @@ const DatePickerExamples = () => (
     <br />
     <h2 id="start-date">Default Date</h2>
     <DatePickerDefaultDate />
+    <br />
+    <br />
+    <h2 id="modal">Presented Inside Modal</h2>
+    <DatePickerInsideModal />
+    <br />
   </div>
 );