+++ packages/terra-modal-manager/tests/nightwatch/modal-manager-spec.js
@@ -15,7 +15,7 @@ module.exports = {
   'Renders the ModalManager': (browser) => {
     browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/modal-manager-tests/demo`);
 
-    browser.expect.element('.terra-ModalManager').to.be.present;
+    browser.expect.element('[class*="container"]').to.be.present;
   },
 
   'Opens the modal when disclose is selected': (browser) => {
@@ -86,23 +86,23 @@ module.exports = {
 
     browser.click('#root-component .disclose');
 
-    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="modal"] [class*="slide-group"] #DemoContainer-1').to.be.present;
 
     browser.click('#DemoContainer-1 .disclose');
 
     // Waiting here to ensure new component is presented and back button is clickable
     browser.waitForElementPresent('[class*="slide-group"] > span > div[class*="slide"]:not([class*="enter-active"]):nth-child(2)', 1000);
 
-    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-1').to.be.present;
-    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-2').to.be.present;
+    browser.expect.element('[class*="modal"] [class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="modal"] [class*="slide-group"] #DemoContainer-2').to.be.present;
     browser.expect.element('[class*="slide-group"] #DemoContainer-2 button.go-back').to.be.present;
 
     browser.click('#DemoContainer-2 button.go-back');
 
     browser.waitForElementNotPresent('#DemoContainer-2', 1000);
 
-    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-1').to.be.present;
-    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-2').to.not.be.present;
+    browser.expect.element('[class*="modal"] [class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="modal"] [class*="slide-group"] #DemoContainer-2').to.not.be.present;
   },
 
   'Releases/Requests focus when selected within modal': (browser) => {