+++ packages/terra-app-delegate/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added requestFocus and releaseFocus attributes to support trapping and releasing focus.
 
 1.1.0 - (July 18, 2017)
 ------------------
@@ -11,9 +13,6 @@ Unreleased
 
 ### Removed
 * Removed nightwatch.config file
-
-### Added
-* Added requestFocus and releaseFocus attributes to support trapping and releasing focus.
 
 1.0.0 - (June 28, 2017)
 ------------------